feat(data table manager): refactor data table manager to move all logic to context
