+++ packages/components/data-table-manager/src/data-table-settings/index.ts
@@ -2,4 +2,5 @@ export {
   default,
   type TDataTableSettingsProps,
   type TSelectChangeEvent,
+  type TColumnManagerProps,
 } from './data-table-settings';
